Fabulous/requirements.txt
#Data collec
yfinance
#sqlite
#data anlaysis stack
datetime
numpy
pandas
statsmodels
#Visualiz stack
matplotlib
seaborn
plotly
#ML Stack
transformers
scikit-learn
tensorflow
keras
xgboost
pandas_ta
ta
TA-Lib
#App Stack
streamlit
streamlit-aggrid
pyfinviz
pytorch
streamlit-autorefresh